refactor(test): split NopSCADlib kernel/brep batches into per-model files
The batched kernel CSG unit tests (nopscad_batch2-5_test.go + nopscad_profile_
test.go) are split one TestNop*CSG per file (nopscad_<part>_test.go), matching the
first models' one-file-per-part convention; batch-local helpers move to
nopscad_batch_helpers_test.go. Pure mechanical move — same tests, kernel/brep green.
