+++ kernel/brep/nopscad_hanginghole_test.go
@@ -1,0 +1,28 @@
+// SPDX-License-Identifier: GPL-2.0-only
+
+package brep_test
+
+import (
+	stdmath "math"
+	"oblikovati/kernel/ops"
+	"oblikovati/math"
+	"testing"
+)
+
+// TestNopHangingHoleCSG pins the printable support volume generated by hanging_hole:
+// a faceted vertical hole column unioned with a square support block below it.
+func TestNopHangingHoleCSG(t *testing.T) {
+	column := prismBody(regularPolygonPoints(math.P3(0, 0, 0), 0.25, 16, 0), 0, 1.4, "hanging-hole-column")
+	support := box(-0.45, -0.45, -0.3, 0.9, 0.9, 0.5)
+	body, err := ops.Boolean(ops.Join, support, column)
+	if err != nil {
+		t.Fatalf("Boolean(Join support+column): %v", err)
+	}
+
+	requireValidNopSolid(t, "hanging_hole", body)
+	columnArea := nopPolygonArea(regularPolygonPoints(math.P3(0, 0, 0), 0.25, 16, 0))
+	want := 0.9*0.9*0.5 + columnArea*1.4 - columnArea*0.2
+	if got := vol(body); stdmath.Abs(got-want)/want > 1e-3 {
+		t.Errorf("hanging_hole volume = %.6f, want %.6f", got, want)
+	}
+}
